requirements.txt
mysql-connector-python==8.0.32
python-dotenv==1.0.0
